+++ src/main/java/racingcar/view/OutputView.java
@@ -1,4 +1,7 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.RacingCar;
+import racingcar.domain.RacingCars;
 
 import java.util.List;
