+++ src/test/java/racingcar/RacingCarTest.java
@@ -3,6 +3,9 @@ package racingcar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RacingCar;
+import racingcar.domain.RacingCars;
+import racingcar.validator.Validator;
 
 import java.util.ArrayList;
 import java.util.Arrays;
